* file.c (rb_file_s_rename): deleted code to get rid of a bug of 	  old Cygwin. 	* file.c (rb_file_truncate): added prototype of GetLastError() 	  on cygwin.  [ruby-dev:31239] 	* intern.h (is_ruby_native_thread): prototype. 	* missing/strftime.c (strftime): fix printf format and actual 	  arguments. 	* ext/Win32API/Win32API.c (Win32API_initialize): ditto. 	* ext/tk/tcltklib.c (ip_finalize): ditto. 	* ext/dl/ptr.c (rb_dlptr_inspect): ditto.  [ruby-dev:31268] 	* ext/dl/sym.c (rb_dlsym_inspect): ditto. 	* ext/socket/getnameinfo.c: include stdio.h always. 	* ext/win32ole/win32ole.c (ole_hresult2msg, folevariable_name, 	  folevariable_ole_type, folevariable_ole_type_detail, 	  folevariable_value, folemethod_visible): missing return value.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/branches/ruby_1_8_6@13200 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ missing/strftime.c
@@ -445,7 +445,7 @@ strftime(char *s, size_t maxsize, const char *format, const struct tm *timeptr)
 			} else {
 				tbuf[0] = '+';
 			}
-			sprintf(tbuf+1, "%02d%02d", off/60, off%60);
+			sprintf(tbuf+1, "%02u%02u", (unsigned)off/60, (unsigned)off%60);
 			break;
 #endif /* MAILHEADER_EXT */
